split out reduction update band after thread mapping
The reduction detection mechanism splits out the band containing the
reduction loop into a separate band in the tree.  As a result, mapping
to threads was made recursive to allow for mapping both the reduction
update band and its parent band node, although mapping nested bands to
threads is invalid in the general case.  However, it is necessary to
split out the reduction loop in order to insert the reduction
synchronization call.

Split out the reduction loop immediately before inserting the reduction
synchronization.  During thread mapping, keep the reduction loop inside
its original band.  This avoids the need to keep track of the relation
between the reduction band and its parent, which can be replaced by
the position of the reduction loop in the band.  As memory promotion may
split the band to insert copies, it may be necessary to traverse the
tree downwards to recover the position of the reduction band member
before inserting the reduction synchronization.