disallow mapping nested bands to threads
mapInnermostBandsToThreads maps coincident band members to threads while
traversing the tree recursively in DFS postorder.  It is possible that
members of two nested bands get mapped to thread dimensions, which was
necessary for parallel reduction support.  The previous commit keeps the
reduction in its original band, removing the need for mapping nested
bands. Such mapping is not valid in general because members of different
bands are not necessarily permutable with each other.  Since thread
dimensions are not executed in any specific order, i.e. there is no
guarantee that, for a given value of threadIdx.y, all threadIdx.x
threads will be executed before (or after) those with another value of
threadIdx.y, permutability is a necessary condition for mapping to
multiple thread dimensions.  Only allow mapping a band to threads if its
children are not already mapped to threads.  The procedure continues
recursively so as to insert all relevant synchronization between
children of sequence nodes even after the thread mapping is complete.